Reduce flickering of highlights while searching
Previously, hardcoded work budgets in terms of loop iterations were
used in clearHighlights, traverseNodes, and timeredHighlightOfMatches.
The problem with these budgets is they're in all likelihood way too low
for most web pages, especially on modern machines.

So, instead of these hardcoded loop iteration budgets, we'll switch to
using a time based budgeting scheme, where we'll get as much work as we
can done in 15ms, and only then if there's still work to do after 15ms
have elapsed will we yield the main thread via the workTimer. Since
clearHighlights can call traverseNodes, and traverseNodes can call
timeredHighlightOfMatches, the max total time we can spend processing
per runloop turn is 45ms.

In practice, this budget seems to be long enough on my machine to allow
the full clear/traverse/highlight process to finish completely every
time, so we get nice flicker free highlighting. But, in the event that
a page is super long and complex, or the machine is running slowly, we
can always still make sure to yield the main thread and defer work to
keep the app responsive.